+++ relevanceai/analytics_funcs.py
@@ -2,8 +2,10 @@ import analytics
 import asyncio
 import json
 import os
-from typing import Callable
+import threading
+import copy
 
+from typing import Callable
 from base64 import b64decode as decode
 from functools import wraps
 
@@ -22,6 +24,13 @@ def get_json_size(json_obj):
     return len(json.dumps(json_obj).encode("utf-8")) / 1024
 
 
+def fire_and_forget(f):
+    def wrapped():
+        threading.Thread(target=f).start()
+
+    return wrapped
+
+
 TRANSIT_ENV_VAR = "_IS_ANALYTICS_IN_TRANSIT"
 
 
@@ -36,9 +45,18 @@ def track(func: Callable):
         try:
             if is_tracking_enabled():
 
-                async def send_analytics():
-                    user_id = args[0].firebase_uid
+                @fire_and_forget
+                def send_analytics():
+                    properties = {}
+                    if "firebase_uid" in kwargs:
+                        user_id = kwargs["firebase_uid"]
+                    elif hasattr(args[0], "firebase_uid"):
+                        user_id = args[0].firebase_uid
+                    else:
+                        user_id = "firebase_uid_not_detected"
+
                     event_name = f"pysdk-{func.__name__}"
+
                     # kwargs.update(dict(zip(func.__code__.co_varnames, args)))
                     # all_kwargs = copy.deepcopy(kwargs)
                     # all_kwargs = all_kwargs.update(
@@ -47,11 +65,29 @@ def track(func: Callable):
 
                     additional_args = dict(zip(func.__code__.co_varnames, args))
 
-                    properties = {
-                        "additional_args": additional_args,
-                        "args": args,
-                        "kwargs": kwargs,
-                    }
+                    if "dataset_id" in kwargs:
+                        properties["dataset_id"] = kwargs["dataset_id"]
+
+                    elif "dataset_id" in additional_args:
+                        properties["dataset_id"] = additional_args["dataset_id"]
+
+                    elif hasattr(args[0], "dataset_id"):
+                        properties["dataset_id"] = args[0].dataset_id
+
+                    elif "self" in additional_args:
+                        if hasattr(additional_args["self"], "dataset_id"):
+                            properties["dataset_id"] = additional_args[
+                                "self"
+                            ].dataset_id
+
+                    full_properties = copy.deepcopy(properties)
+                    full_properties.update(
+                        {
+                            "additional_args": additional_args,
+                            "args": args,
+                            "kwargs": kwargs,
+                        }
+                    )
                     if user_id is not None:
                         # TODO: Loop through the properties and remove anything
                         # greater than 5kb
@@ -62,23 +98,26 @@ def track(func: Callable):
                             or "fit" in event_name
                             or "predict" in event_name
                             or get_json_size(
-                                json_encoder(properties, force_string=True)
+                                json_encoder(full_properties, force_string=True)
                             )
                             > 30
                         ):
                             response = analytics.track(
                                 user_id=user_id,
                                 event=event_name,
+                                properties=json_encoder(properties, force_string=True),
                             )
                         else:
                             response = analytics.track(
                                 user_id=user_id,
                                 event=event_name,
-                                properties=json_encoder(properties, force_string=True),
+                                properties=json_encoder(
+                                    full_properties, force_string=True
+                                ),
                             )
 
-                # send_analytics()
-                asyncio.ensure_future(send_analytics())
+                send_analytics()
+                # asyncio.ensure_future(send_analytics())
         except Exception as e:
             pass
         try: